requirements.txt
pandas
numpy<2.0
torch
torchvision
Pillow
scikit-learn
matplotlib
streamlit